% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deota}
\alias{deota}
\title{Phases of clock genes}
\format{
An object of class \code{list} of length 2.
}
\source{
\href{https://doi.org/10.1016/j.cmet.2022.12.006}
}
\usage{
data(deota)
}
\description{
Data from an experiment performed on mice to determine compare
the peak phase times of circadian clock genes under ad-libitum
and time-restricted feeding regimines. The two imported data frames
are \code{alf} and \code{trf} corresponding to ad-libitum and time-restricted
results, respectively. The observations are measured in degrees,
corresponding to time on a 24-hour zeigeber time clock cycle.
}
\references{
Deota S, Lin T, Chaix A, Williams A, Le H, Calligara H,
Ramasamy R, Huang L, Panda S (2023) Diurnal transcriptome landscape
of a multi-tissue response to time-restricted feeding in mammals.
Cell Metabolism 35(1):150-164.e4
}
\keyword{datasets}
