% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupedCircularBoxplot.R
\name{GroupedCircularBoxplot}
\alias{GroupedCircularBoxplot}
\title{Create grouped circular boxplot}
\usage{
GroupedCircularBoxplot(
  data_in,
  template = "degrees",
  place = "none",
  units = "degrees",
  marg = "large",
  shrink = 1.6,
  H = FALSE,
  stack = FALSE,
  constant = "optimal",
  rad_shift = 0.5,
  plot_cols = RColorBrewer::brewer.pal(8, "Set2"),
  line_cols = RColorBrewer::brewer.pal(8, "Dark2"),
  col_type = "fill",
  legend_pos = "topleft",
  draw_arrow = TRUE,
  ordinal = FALSE
)
}
\arguments{
\item{data_in}{List of circular objects to be plotted}

\item{template}{One of "degrees", "radians", "geographics", or NULL}

\item{place}{If template is NULL, either "outside" or "inside" denoting where the axis should be drawn}

\item{units}{If template is NULL, units to use on the drawn axis}

\item{shrink}{TODO}

\item{H}{Logical indicating if each data point should be drawn in addition to the boxplot}

\item{stack}{Logical indicating if drawn points should be stacked}

\item{constant}{TODO}

\item{rad_shift}{Scalar indicating the distance between each boxplot}

\item{plot_cols}{Vector with the same length as \code{data_in}, specifying the color of the boxplot}

\item{line_cols}{Vector with the same length as \code{data_in}, specifying the color of the median lines and arrows}

\item{col_type}{String which doesn't do anything I thing}

\item{legend_pos}{String indicating where the legend should be drawn, or "none" for no legend}

\item{draw_arrow}{Logical specifying if arrows pointing to each median should be drawn}

\item{ordinal}{Logical, if \code{template} is NULL and \code{units} is "geographics", should the ordinal directions also be drawn?}
}
\description{
does some stuff
}
\author{
Josh Berlinski
}
